category list - broken?

+++ CMS/PDF/PDF.php
@@ -5,8 +5,7 @@
  * Date: 12/04/16
  * Time: 16:07
  */
-use dbconnect;
-//dbconnect\dbconfig::connect();
+//use dbconnect\dbconnect;
 
 ini_set('display_errors', 'On');
 error_reporting(E_ALL);
@@ -36,19 +35,20 @@ error_reporting(E_ALL);
     }
 
 //set variables used for file upload
-$target_dir = "";
-$target_file ="";
-$target_file = $target_dir . basename($_FILES["UpPDF"]["name"]);
+$targetDir = "";
+$targetFile ="";
+$targetFile = $targetDir . basename($_FILES["UpPDF"]["name"]);
+
 $uploadOk = 1;
-$fileType = pathinfo($target_file,PATHINFO_EXTENSION);
+$fileType = pathinfo($targetFile,PATHINFO_EXTENSION);
 
     class PDF
     {
 
 
         public $Category;
-        protected $target_dir;
-        protected $target_file;
+        protected $targetDir;
+        protected $targetFile;
         protected $uploadOk;
         protected $fileType;
         protected $targetNospace;
@@ -59,12 +59,12 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
     public static function upload()
     {
 
-        global $target_dir, $target_file , $uploadOk ,$fileType ,$targetNospace;
+        global $targetDir, $targetFile , $uploadOk ,$fileType ,$targetNospace;
 
-        $targetNospace = str_replace(' ', '_', $target_file);
+        $targetNospace = str_replace(' ', '_', $targetFile);
 
         // Check if file already exists
-        if (file_exists($target_file)) {
+        if (file_exists($targetFile)) {
             echo "Sorry, file already exists.";
             $uploadOk = 0;
 
@@ -119,12 +119,12 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
 //***********************************************************************************************************************
         public static function insert()
         {
-            global $target_dir, $target_file , $uploadOk ,$fileType;
+            global $targetDir, $targetFile , $fileName;
 
             //new instance of CatarrayF function
             $catArray = CatArrayF();
 
-            $UploadFileName = $_FILES['UpPDF']['name'];
+          //  $UploadFileName = $_FILES['UpPDF']['name'];
             $catSelect = $_POST['Cat'];
             echo"<br>";
             echo"<br>";
@@ -132,15 +132,21 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
             //count
             for ($i = 0; $i < 10; $i++) {
 
-                $CatSear = array_search($catSelect, $catArray[$i]);
+                $catSear = array_search($catSelect, $catArray[$i]);
 
-                if($CatSear == true){
+                var_dump($catSear);
+
+                echo "<br>";
+                if($catSear >= 0){
+                    echo "ello";
+
 
                     //get array value (ie:"SSOF") from index number
-                    $category = $catArray[$i][0];
+                   $category = $catArray[$i][0];
                     //set file path to place file
-                    $target_dir = "Category/".$category ."/";
-                    $target_file = $target_dir . basename($_FILES["UpPDF"]["name"]);
+                    $targetDir = "Category/".$category ."/";
+                    $targetFile = $targetDir . basename($_FILES["UpPDF"]["name"]);
+                    $fileName = basename($_FILES["UpPDF"]["name"]);
                     //set post values for DB
                     $title = $_POST['Title'];
                     $info = $_POST['Info'];
@@ -150,33 +156,21 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
                     $fileExt= $FE->getExtension();
 
                     //regex : remove extension and replace spaces with underscores
-                    $targetWithoutExt = preg_replace('/\\.[^.\\s]{3,4}$/', '', $target_file);
+                    $targetWithoutExt = preg_replace('/\\.[^.\\s]{3,4}$/', '', $targetFile);
                     $targetWithoutExt = str_replace(' ', '_', $targetWithoutExt);
 
 
-                    //code here **********************************
 
-
-                    //connect to DB
-
-                    dbconnect\dbconfig::connect();
-
-                    //$conn = new dbconfig();
-                    //$conn = $conn->connect();
 
                     $sql= "";
 
 
                     try {
+                        //connect to DB
                         $conn = new PDO("mysql:host=localhost;dbname=Pestproof", 'baine101', 'blink182');
                         // set the PDO error mode to exception
                         $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                        $sql = "INSERT INTO PDF (Title, Info, Cat, Path, Filetype) VALUES ( '$title', '$info', '$category', '$targetWithoutExt', '$fileExt')";
-                        //$sql->bindParam(':title', $title);
-                        //$sql->bindParam(':info', $info);
-                        //$sql->bindParam(':category', $category);
-                        //$sql->bindParam(':targetWithoutExt', $targetWithoutExt);
-                        //$sql->bindParam(':fileExt', $fileExt);
+                        $sql = "INSERT INTO PDF (Title, Info, Cat, Path, FileName, Filetype) VALUES ( '$title', '$info', '$category', '$targetWithoutExt','$fileName' , '$fileExt')";
 
                         if(PDF::upload() == true) {
 
@@ -197,6 +191,7 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
                 }
                //close for $a1
                }
+
 
             //return true;
         //close Upload Func
@@ -226,66 +221,91 @@ $fileType = pathinfo($target_file,PATHINFO_EXTENSION);
             //include "../dbConfig.php";
 
             $catArray = CatArrayF();
-
+            //a1 = cycle through outer array
             for ($a1 = 0; $a1 < 10; $a1++)
             {
-
-
+                //echo each category HTML
                 echo"<h2>".$catArray[$a1][1]."</h2>";
-
-
 
                  //count the vars in inner array
                 for ($row = 1; $row < 2; $row++)
             {
-
-               $ID = $catArray[$a1][0];
-                $value = $catArray[$a1][1];
-
-
-                //connect to DB
-                //dbconnect\dbconfig::connect();
-
                 $conn="";
-
 
                 try {
                     $conn = new PDO("mysql:host=localhost;dbname=Pestproof", 'baine101', 'blink182');
                     // set the PDO error mode to exception
                     $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
-                    $result = $conn->query("SELECT * FROM PDF WHERE Cat = '$ID'");
+                    $ID = $catArray[$a1][0];
+                    $value = $catArray[$a1][1];
 
-                    var_dump($result);
+
+                    //count all the files in the category/directory
+                    $sqlCount= "SELECT COUNT(*) FROM PDF WHERE Cat = '$ID'";
+                    $sqlRowCount =  $conn->query($sqlCount);
+
+                    //if their is a row in the DB then :
+                    if ($sqlRowCount->rowCount() > 0) {
+                        var_dump($sqlRowCount->rowCount());
+
+                        //real sql query that return all values in DB as array
+                        $sql= "SELECT * FROM PDF WHERE Cat = '$ID'";
+                        $result =  $conn->query($sql);
+
+                        // output data of each row
+
+                        while($row = $result->fetchAll(PDO:: FETCH_ASSOC))
+                        {
+                            global $str;
+                            //loop through the files found
+                            for($count = 0; $count <15 ; $count++)
+                            {
+
+                                if(isset($row[$count]['FileName'])) {
+                                    //set directory path to check if file exists
+                                    $dir = "PDF/Category/" . $ID . "/";
+                                    $targetFile = str_replace(' ', '_', $row[$count]['FileName']);
+                                    $str = $dir . $targetFile;
+                                }
+
+
+                                //if the ID and Title colums are not empty
+                                if(!empty($row[$count]["ID"]) && !empty($row[$count]["Title"]) && file_exists($str) )
+                                {
+
+                                        //display files HTML
+                                        echo "id : " . $row[$count]["ID"] . "<br> Name : " . $row[$count]["Title"] . "<br> Info : " . $row[$count]["Info"] . "<br> Category : " . $row[$count]['Cat'] . "<br> path : " . $row[$count]['Path'] . "<br> ext: " . $row[$count]['FileType'] . "<br> end";
+                                        echo "<br>";
+                                        echo "<br>";
+
+
+                                 //!!!!!!!!!!!code stops here when its not suposed to :@ !!!!!!!!!!!!
+
+
+                                //close if isset $row[$count]["ID"]
+                                }
+                            //close for count
+                            }
+
+                        //close while $row
+                        }
+                    } else {
+                        echo "0 results";
+                    }
 
                 }catch(PDOException $e)
                     {
                     echo $conn . "<br>fail:" . $e->getMessage();
                     }
 
-
-
             // close for row
             }
-
-
-                var_dump($ID);
-                echo"<br> Value:";
-                var_dump($value);
-                echo"<br>";
-
-
-
-
-
             //close for a1
             }
 
-
-
+        //close function PDFList
         }
-
-
 
 //close class PDF
     }
@@ -297,8 +317,8 @@ $PDF = new PDF;
 
 if(isset($_POST['Upload'])){
 
-    //PDF::insert();
-    PDF::PDFList();
+    PDF::insert();
+    //PDF::PDFList();
 }
